# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/Cart/checkout/Cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Show Products in Cart",
                "operationId": "688199ce3b77a3635bc5edfc9c6d6ebe",
                "responses": {
                    "200": {
                        "description": "Data index successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/Cart/Add/Product": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add Product in Cart",
                "operationId": "d45f5d1868d27e9350cf7f60e2477a6e",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "id product",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "color_id",
                        "in": "query",
                        "description": "id product",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data index successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Cart/Added/Product": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add Added in Cart",
                "operationId": "7d370dc4235b29b97fbabb204babb4b0",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "id product",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "color_id",
                        "in": "query",
                        "description": "id product",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data index successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Cart/Decreased/Product": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Decreased Product in Cart",
                "operationId": "dbcf1b97a04de46c0d4f6bc0337cf990",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "id product",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "color_id",
                        "in": "query",
                        "description": "id product",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data index successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Product/index": {
            "get": {
                "tags": [
                    "Product Home"
                ],
                "summary": "index Products",
                "operationId": "5e61ecf4d3fbfff0117b154169396340",
                "responses": {
                    "200": {
                        "description": "Data create successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                }
            }
        },
        "/api/Product/Show": {
            "get": {
                "tags": [
                    "Product Home"
                ],
                "summary": "index Products",
                "operationId": "76b5ef21832f71911d455531fc28518a",
                "parameters": [
                    {
                        "name": "dtp",
                        "in": "query",
                        "description": "dtp product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data create successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}